Sidebar UIの改善とスタイリング
- SideBarコンポーネントを新しいディレクトリに移動
- Tailwind CSSを使用してサイドバーのレイアウトとスタイルを強化
- リンクとユーザー情報のスタイリングを追加
- レスポンシブデザインとユーザビリティの向上
